chsysroot.sh makes an executable using libraries installed in a directory.
The build time libraries may not the same as the target system and we
don't want to install our libraries to the system directories.  In
order to install our library in an independent directory, we set rpath
and interpreter for executsbles built by us.

The script supposes that we copy our libraries to <sysroot>/lib and
<sysroot>/usr/lib on the target system. It will add these two paths to
rpath of the executable to make sure it uses libraries under these two
directories. It also sets the interpreter path to the RTLD in
<sysroot>/lib.