ports: run chrootsys.sh on all binaries listed in the BINS var.
The caller has to provide DEST_SYSROOT that is the path that binaries
and libraries will be installed on the target machine. The path is
where toolchain/ will be on the target machine. For example,
toolchain/ will be /path/to/myroot/ on the target machine, then
DEST_SYSROOT should be /path/to/myroot.

For example, in a port directory:

    make DEST_SYSROOT=/path/to/myroot initsysroot

Or, you defined DEST_SYSROOT in dest-config.mk at the source root
directory of this build system.